third_party/WebKit/LayoutTests/virtual/off-main-thread-fetch/http/tests/fetch/workers/block-mixed-content-base-https-expected.txt

CONSOLE ERROR: line 157: Fetch API cannot load http://example.test:8000/serviceworker/resources/fetch-access-control.php?ACAOrigin=*&label=test1-same-origin. Request mode is "same-origin" but the URL's origin is not same as the request origin https://127.0.0.1:8443.
CONSOLE ERROR: Fetch API cannot load https://127.0.0.1:8443/serviceworker/resources/redirect.php?ACAOrigin=*&Redirect=http%3A%2F%2Fexample.test%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*%26label%3Dtest2-same-origin. Redirect failed.
CONSOLE ERROR: line 170: Fetch API cannot load http://example.test:8000/serviceworker/resources/redirect.php?ACAOrigin=*&Redirect=https%3A%2F%2F127.0.0.1%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*%26label%3Dtest3-same-origin. Request mode is "same-origin" but the URL's origin is not same as the request origin https://127.0.0.1:8443.
CONSOLE ERROR: line 188: Fetch API cannot load https://localhost:8443/serviceworker/resources/redirect.php?ACAOrigin=*&Redirect=https%3A%2F%2F127.0.0.1%3A8443%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*%26label%3Dtest5-same-origin. Request mode is "same-origin" but the URL's origin is not same as the request origin https://127.0.0.1:8443.
This is a testharness.js-based test.
PASS Startup 
PASS Mixed content fetch (same-origin, HTTPS->HTTP) 
PASS Mixed content redirect (same-origin, HTTPS->HTTPS->HTTP) 
PASS Mixed content redirect same-origin, HTTPS->HTTP->HTTPS) 
PASS Same origin redirect (same-origin, HTTPS->HTTPS->HTTPS) 
PASS Cross origin redirect (same-origin, HTTPS->HTTPS->HTTPS 
FAIL Mixed content fetch (cors, HTTPS->HTTP) assert_unreached: unexpected fulfillment: [object Response] Reached unreachable code
FAIL Mixed content redirect (cors, HTTPS->HTTPS->HTTP) assert_unreached: unexpected fulfillment: [object Response] Reached unreachable code
FAIL Mixed content redirect cors, HTTPS->HTTP->HTTPS) assert_unreached: unexpected fulfillment: [object Response] Reached unreachable code
PASS Same origin redirect (cors, HTTPS->HTTPS->HTTPS) 
PASS Cross origin redirect (cors, HTTPS->HTTPS->HTTPS 
Harness: the test ran to completion.
